pyproject.toml
[project]
name = "trmnl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "pillow>=10.0.0",
]
